# Optimize mapSubcript by caching input keys in a hash map.