--- a/velox/functions/prestosql/registration/GeneralFunctionsRegistration.cpp
+++ b/velox/functions/prestosql/registration/GeneralFunctionsRegistration.cpp
@@ -19,6 +19,12 @@
 #include "velox/functions/prestosql/Cardinality.h"
 
 namespace facebook::velox::functions {
+extern void registerSubscriptFunction(
+    const std::string& name,
+    bool enableCaching);
+extern void registerElementAtFunction(
+    const std::string& name,
+    bool enableCaching);
 
 // Special form functions don't have any prefix.
 void registerAllSpecialFormGeneralFunctions() {
@@ -29,8 +35,9 @@ void registerAllSpecialFormGeneralFunctions() {
 }
 
 void registerGeneralFunctions(const std::string& prefix) {
-  VELOX_REGISTER_VECTOR_FUNCTION(udf_element_at, prefix + "element_at");
-  VELOX_REGISTER_VECTOR_FUNCTION(udf_subscript, prefix + "subscript");
+  registerSubscriptFunction(prefix + "subscript", true);
+  registerElementAtFunction(prefix + "element_at", true);
+
   VELOX_REGISTER_VECTOR_FUNCTION(udf_transform, prefix + "transform");
   VELOX_REGISTER_VECTOR_FUNCTION(udf_reduce, prefix + "reduce");
   VELOX_REGISTER_VECTOR_FUNCTION(udf_array_filter, prefix + "filter");