… velox/functions/prestosql/benchmarks/CMakeLists.txt
… velox/functions/prestosql/benchmarks/MapSubscriptCachingBenchmark.cpp
… velox/functions/prestosql/tests/ElementAtTest.cpp
… velox/functions/sparksql/Register.cpp