Summary:
There are cases where the mapSubscript function can receive the same (base) map over and over.
In that case searching the same map over and over is redundant.

This optimization makes mapSubscript a stateful function, if it sees that the same
base map is being provided for it, it will cache a materialized version of the input.

The shadowed query e2e runtime goes from 1.50 hours reduce to 16.48 min.
a local benchmark is added, the optimization is only enabled for non-bool pritmive types.

A benchmark is added. The function it self become so much faster speedups depends on the
number of batches, and the size of the base vector. The production case we have 1 map of size 80k
entry! so speedups for the function is extremely high.
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
INTEGER_10000_1##subscript                                 29.09ms     34.37
INTEGER_10000_1##subscriptNoCashing                          3.08s   324.47m
INTEGER_10000_1000##subscript                              81.38ms     12.29
INTEGER_10000_1000##subscriptNoCashing                       3.19s   313.51m
INTEGER_1000_1##subscript                                  26.94ms     37.12
INTEGER_1000_1##subscriptNoCashing                        325.23ms      3.07
INTEGER_1000_1000##subscript                               40.22ms     24.86
INTEGER_1000_1000##subscriptNoCashing                     356.80ms      2.80
INTEGER_100_1##subscript                                   29.65ms     33.72
INTEGER_100_1##subscriptNoCashing                          51.72ms     19.34
INTEGER_100_1000##subscript                                32.03ms     31.22
INTEGER_100_1000##subscriptNoCashing                       56.16ms     17.81
INTEGER_10_1##subscript                                    23.61ms     42.35
INTEGER_10_1##subscriptNoCashing                           23.89ms     41.86
INTEGER_10_1000##subscript                                 24.46ms     40.88
INTEGER_10_1000##subscriptNoCashing                        24.65ms     40.58
VARCHAR_10000_1##subscript                                 77.48ms     12.91
VARCHAR_10000_1##subscriptNoCashing                          4.60s   217.57m
VARCHAR_10000_1000##subscript                             175.11ms      5.71
VARCHAR_10000_1000##subscriptNoCashing                       8.34s   119.88m
VARCHAR_1000_1##subscript                                  69.91ms     14.30
VARCHAR_1000_1##subscriptNoCashing                        592.40ms      1.69
VARCHAR_1000_1000##subscript                              133.53ms      7.49
VARCHAR_1000_1000##subscriptNoCashing                     617.93ms      1.62
VARCHAR_100_1##subscript                                   63.33ms     15.79
VARCHAR_100_1##subscriptNoCashing                          73.29ms     13.64
VARCHAR_100_1000##subscript                                84.33ms     11.86
VARCHAR_100_1000##subscriptNoCashing                       87.19ms     11.47
VARCHAR_10_1##subscript                                    31.01ms     32.25
VARCHAR_10_1##subscriptNoCashing                           31.62ms     31.62
VARCHAR_10_1000##subscript                                 33.86ms     29.54
VARCHAR_10_1000##subscriptNoCashing                        33.84ms     29.55
```

Differential Revision: D50544250